Shortened logged messages to prevent endline characters causing a crash.
for some god forsaken fucking reason the way embeds in discord count endline characters is not the same as the way javascript counts endline characters, and it leads to the endline character taking up 2 character length in embeds.
man whoever wrote that is fucking retarded and I want to strangle them.